modal scss 수정
