+++ public/data/detail.json
@@ -26,11 +26,6 @@
         "img": "images/products/blacktea.png",
         "title": "블랙티",
         "desc": "살짝 쌉싸래함 뒤 깔끔함"
-      },
-      {
-        "img": "images/products/blacktea.png",
-        "title": "블랙티",
-        "desc": "살짝 쌉싸래함 뒤 깔끔함"
       }
     ]
   }
