+++ src/pages/Detail/Modal/Modal.js
@@ -13,8 +13,8 @@ const Modal = ({ modalOpen, closeModal }, props) => {
               </button>
             </header>
             <main className="modalMain">
-              <h1>상품이 장바구니에 담겼습니다.</h1>
-              <p>바로 확인하시겠습니까?</p>
+              <h1 className="mainFirstLine">상품이 장바구니에 담겼습니다.</h1>
+              <p className="mainSecondLine">바로 확인하시겠습니까?</p>
             </main>
             <footer className="modalFooter">
               <button className="footerCartBtn" onClick={closeModal}>
