+++ src/pages/Detail/Modal/Modal.scss
@@ -1,71 +1,90 @@
 .modal {
   display: none;
   position: fixed;
-
   top: 0;
   right: 0;
   bottom: 0;
   left: 0;
   z-index: 99;
   background-color: rgba(0, 0, 0, 0.6);
-}
-.modalBox {
-  margin: 28px 20px;
+
+  .modalSection {
+    width: 90%;
+    max-width: 450px;
+    margin: 0 auto;
+
+    background-color: #fff;
+    /* 팝업이 열릴때 스르륵 열리는 효과 */
+    animation: modal-show 0.3s;
+    overflow: hidden;
+
+    .modalBox {
+      margin: 28px 20px;
+
+      .modalHeader {
+        position: relative;
+        padding: 16px 64px 16px 16px;
+        background-color: #fff;
+        font-weight: 700;
+
+        .headerCloseBtn {
+          position: absolute;
+          top: 15px;
+          right: 15px;
+          width: 30px;
+          font-size: 35px;
+          font-weight: 300;
+          text-align: center;
+          color: #6c757d;
+          background-color: transparent;
+        }
+      }
+      .modalMain {
+        padding: 16px;
+        text-align: center;
+
+        .mainFirstLine {
+          padding-top: 20px;
+          font-size: 22px;
+          font-weight: 700;
+          padding-bottom: 15px;
+        }
+        .mainSecondLine {
+          font-size: 18px;
+          background: none;
+          padding-bottom: 20px;
+        }
+      }
+
+      .modalFooter {
+        padding: 12px 16px;
+        text-align: center;
+
+        .footerCartBtn {
+          padding: 18px 35px;
+          margin-right: 20px;
+          color: #000;
+          background-color: #fff;
+          font-size: 13px;
+          outline: none;
+          cursor: pointer;
+          border: 1px solid #000;
+        }
+
+        .footerOrderBtn {
+          padding: 18px 45px;
+          color: #fff;
+          background-color: #000;
+          font-size: 13px;
+          outline: none;
+          cursor: pointer;
+          border: 0;
+        }
+      }
+    }
+  }
 }
 
-.modalSection {
-  width: 90%;
-  max-width: 450px;
-  margin: 0 auto;
-
-  background-color: #fff;
-  /* 팝업이 열릴때 스르륵 열리는 효과 */
-  animation: modal-show 0.3s;
-  overflow: hidden;
-}
-.modalHeader {
-  position: relative;
-  padding: 16px 64px 16px 16px;
-  background-color: #fff;
-  font-weight: 700;
-}
-.headerCloseBtn {
-  position: absolute;
-  top: 15px;
-  right: 15px;
-  width: 30px;
-  font-size: 35px;
-  font-weight: 300;
-  text-align: center;
-  color: #6c757d;
-  background-color: transparent;
-}
-.modalMain {
-  padding: 16px;
-}
-.modalFooter {
-  padding: 12px 16px;
-
-  text-align: center;
-}
-.footerCartBtn {
-  padding: 6px 12px;
-  color: #fff;
-  background-color: #6c757d;
-  font-size: 13px;
-  outline: none;
-  cursor: pointer;
-  border: 0;
-}
-.footerOrderBtn {
-  padding: 6px 12px;
-  color: #fff;
-  background-color: #6c757d;
-  font-size: 13px;
-  outline: none;
-  cursor: pointer;
-  border: 0;
-}
 .modal.openModal {
   display: flex;
   align-items: center;
