+++ src/pages/Detail/Tab/Tab.scss
@@ -1,4 +1,6 @@
 .tab {
+  position: relative;
+  z-index: 0;
   scroll-behavior: smooth;
 
   .tabNav {
@@ -14,7 +16,7 @@
       width: 100%;
       text-align: center;
 
-      a {
+      .info {
         position: relative;
         display: block;
         background: #ffffff;
@@ -35,13 +37,7 @@
           width: 100%;
           height: 3px;
         }
-      }
-    }
-  }
 
-  .tabNav {
-    li {
-      a {
         &.active {
           border-left: 1px solid #111;
           border-right: 1px solid #111;
